+++ packages/nns/src/canisters/governance/request.converters.spec.ts
@@ -1,6 +1,6 @@
 import { Principal } from "@dfinity/principal";
 import { arrayBufferToUint8Array, toNullable } from "@dfinity/utils";
-import type { ManageNeuron as RawManageNeuron } from "../../../candid/governance";
+import type { ManageNeuronRequest as RawManageNeuron } from "../../../candid/governance";
 import {
   CanisterAction,
   CanisterInstallMode,